Moved the ClientApp and ServerApp into fed_app.py.
